pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "kulms-download"
description = "Download site resources from KULMS, Kyoto University Learning Management System."
readme = { file = "README.md", content-type = "text/markdown" }
dynamic = ["version"]
requires-python = ">= 3.12"
dependencies = [
    "aiofiles",
    "httpx",
    "keyring",
    "platformdirs",
    "pywebview",
    "questionary",
    "rich",
]
keywords = ["kyoto", "ku", "download", "kulms"]

[project.scripts]
kulms-download = "kulms_download.main:main"

[project.urls]
Repository = "https://github.com/" # TODO: ここをちゃんと書く

[tool.setuptools_scm]
